requirements.txt
python-telegram-bot
python-dotenv
SQLAlchemy
cryptography
pybit
telethon